PATCH: [perl #75138] "\c`" -> " "
Attached is a patch for some of this issue.  I took Nicholas' advice,
and if the result of \cX isn't a word character, the output message will
precede it with a backslash, so the message in the example would be

"\c`" more clearly written simply as "\ " at -e line 1.

I think that message is true.

I also added tests.

There is a test that guarantees that we won't ship 5.14 with things as
they are now in it.  I added wording to the comments next to that test
to be sure to verify with this email thread if we should remove the
deprecation, and mentioned that in the explanatory wording in the pod.
I support removing the deprecation, but for now I'm not touching that,
to see what other issues may yet arise before 5.14.