perlbug: always allow change of sender address
The code that guessed a default sender address would under some
circumstances not display this guess to the user with the option to change
it. Specifically, if $Config{cf_me} eq {login id of the user running
perlbug}.

In my case, cf_email got set at build time (Configure default) of
'davem@pigeon.(none)', which was then silently used. Not good!