Avoid creating @EXPORT_FAIL in every package using Exporter.
Previously, if package INKLE_KLINK is an Exporter, then the Exporter code's
symbolic lookup of \@{"INKLE_KLINK::EXPORT_FAIL"} would cause both that array
and the enclosing typeglob to be instantiated. Now the typeglob and array are
only created if present. (Strictly, if there's something in INKLE_KLINK's
symbol table for EXPORT_FAIL. But likely that will only ever be the array.)

This saves about 200 bytes per package that uses Exporter but does not need an
@EXPORT_FAIL.