In Config::import, avoid assigning to $pkg, which is never read.
Avoid the use of sprintf - __PACKAGE__ is a constant, so can be inlined within
the message string. Don't store '%Config' in %Export_Cache, as that's only used
to deal with exported functions. The variable %Config is special-cased.

On this platform, cachegrind shows a 1% reduction in instruction count and
L1 cache references as a result.