Add a tool for writing a perldelta using git notes
I've written this tool to help me annotate commits for perldelta.
It scans back through commits, prompting for a perldelta section
for each commit (or to ignore the commit).  It then opens up an
editor to write a perldelta snippet into a git note on the commit.

It also supports some very primative workflow, including reviewing
existing annotations and "rendering" annotations properly grouped
by section to cut/paste into perldelta.

It does have some non-core dependencies including Git::Wrapper and
Term::ReadKey, so is intended to be run using an installed perl with
these module from CPAN, not the freshly built one.

Documentation is sparse.

N.B. Git notes are *local* -- they are kept in a detached branch and
will not be pushed upstream.  This makes them well-suited for a
release manager to keep working notes (as for perldelta) that will
become irrelevant over time.  It's not clear whether they have broader
utility.