Refactor loops in S_hsplit(), Perl_hv_ksplit() and Perl_ptr_table_split().
Change from for() to do ... while() loops. Move variable initialisation to
variable declaration. Avoid needing to use the comma operator to allow
multiple statements in for(). Avoid using a continue statement where it
isn't actually needed to change flow control. Avoid relying on the optimiser
to know that the for loop conditional doesn't need testing on the first pass.

A current gcc's optimiser produces identical code despite these changes.
However, for the reasons given I consider the code to be much clearer.