Remove alloca() use from Time::Piece for portability.
In 1.17, T::P started using its own strptime (almost) everywhere,
and this _strptime contains a single reference to alloca().  The
Linux and BSD man pages recommend against using alloca(), the
function  doesn't exist on HP-UX or VMS, and it's impossible to
error check it.  So replace it here with malloc().

If getting three or four bytes from the heap instead of the stack
is considered a problem, an alternative would be researching the
maximum length of a timezone name and using an automatic variable
to hold it.