Store xio_ifp in sv_u in the SV head, reducing XPVIO by 1 pointer.
When accessing a file handle for reading, this reduces pointer dereferences by
1, which will reduce CPU cache pressure.

As SVt_PVIO is also the default type provided to source filters, the code needs
to allow them to continue to use the sv_u for SvPVX(). Re-use the existing
IOf_FAKE_DIRP to signal this, as it's only set when a source filter is added.