charnames: Fix scoping bugs
This was done by moving what could to %^H.  Because data structures in
%^H get stringified at runtime, new serialized entries for them had to
be created and then unserialized on each runtime call.  Also, because
%^H is read-only at runtime, some data structures couldn't be moved to
it.  Things were set up so that these contain only things invariant
under scoping, and looked at only when the same scoped options are in
effect as when they were created.  Further comments at declaration of
%full_names_cache.

I was well into this patch when it dawned on me that it was doing
unnecessary tests, so that
    if (! a) { conditionally set a }
    if (! a) {}

could be implemented more efficiently as

    if (! a) {
	conditionally set a }
	if (! a) {}
    }

so I changed it, which messes up leading indentation for the diffs.