charnames.pm: More refactoring for performance
I realized after the last commit that it might be faster to use a trie
when there are multiple scripts that a letter could be in, instead of
searching the table for each script.  When there were 6 possible scripts
and the letter was found in the final one, the speed-up was a factor of
5.  This also simplified things.  The list of scripts can be stored as a
string like A|B|C instead of a stringified array, and the code just gets
simpler.

Also, there were complications to the code to keep from zapping the
input name, just in case it was needed for an error message.  But I
realized that instead of using a shift to get the name, just copy it
from $_[0], and on the error leg that needs the original, it still is in
$_[0].  If a user-defined alias is to a character name to lookup and
that one is invalid, we want to output the invalid one, so a further
variable, $save_input is used to hold it