[perl 71764] Extend charnames to all of Unicode
This patch causes \N{}, vianame, and viacode to know the names of all
Unicode code points.  Previously the names that are algorithmically
determinable were not handled.  These include the Hangul syllables and
many CJK characters.

It simply adds using the routines that mktables inserts into Name.pl
that handle these characters.  mktables generates these algorithms from
data in the Unicode data base.  The routines have been there since
11/2009 in anticipation of this change, but have been unused until now.
They probably have not been reviewed thoroughly.

The major change to this is the .t file.  Now that all code points are
understood, the .t tests them all.  But this would take too long each
time, so it tests a random sample.  If there is a failure, the seed is
output so that the test can be reproduced.  This idea came from Michael
Schwern, and is the same he uses in Test::Sims.  Various parameters
about the sampling are easily adjustable.