Refactor common parts of op.h, regexp.h into new .h
op.h and regexp.h share common elements in their data structures.  They
have had to manually be kept in sync.  This patch makes it easier by
putting those common parts into a common header #included by the two.

To do this, it seemed easiest to change the symbol definitions to use
left shifts to generate the flag bits.  But this meant that regcomp.pl
and axt/B/defsubs_h.PL had to be taught to recognize those forms of
expressions, done in separate commits