C++ compiler fix for util.c
Even with all of the changes, we still die on a strchr() call now
because glibc provides "correct" strchr prototypes rather than the
C ones.

C:
  char *strchr(const char *s, int c)
C++:
  const char *strchr(const char *s, int c)
  char *strchr(      char *s, int c)

and of course C++ doesn't let you convert a 'const char *' to a
'char *' so boom on util.c:3972 in Perl_grok_bslash_o (due to 'e').