Don't pretend to support really old VC++ compilers
The oldest supported Visual C++ compiler is VC6 (released in 1998),
and we already jump through some hoops for it by inlining newer
API definitions in win32/win32.c.  It is desirable to keep supporting
VC6 because it is the last MSVC release that links against MSVCRT.dll
and not against a compiler specific runtime library.