Global executable bit cleanup
When porting/makerel runs, all files copied into the directory for the
tarball have the executable bit stripped and then only a specific set of
files have the executable bit restored.

There are many files in the repo that have the executable bit set in the
repo that will be stripped.  So that the state of files in the repo is
as close as possible to the state of files in the release tarball, the
executable bit has been stripped from such files.

In one recent case, a file added from a dual-life module needed the
executable bit set.  Because it had the bit in the repo but was
not listed in makerel to get an executable bit, tests using it
passed in the repo and failed in the tarball.

This commit refactors the list into a new file, Porting/exec-bit.txt
and add tests to detect a mismatch between files listed there
and actual executable bits in the repo.