charnames.t: tweak amount of testing of CJK chars
Actually, this tweaks the amount of testing of characters whose names are
algorithmically determinable, most of which are CJK characters.

This patch changes the testing to test not 1% of them, but to test 1 in
each block, no matter what the block size.  We really don't need to test
many of these to be confident the algorithm is working.

It also adds some comments to clarify what happens if one tweaks the
block size.