Remove compile-time conditionals from PVIV and PVNV body sizes.
Alignment issues that aggressive (but legal) sparc compilers tripped over
were resolved by 69ba284b5e077075. Hence the conditional sizing added in
889d28b2ea2c1751 is not needed, and prevents viable size optimisations.

(The compiler could "see" that the structure contained a double, hence the
alignment was constrained, hence it was legal to use a single 64 bit load on 2
adjacent 32 bit values. Which fails badly when the structure isn't actually
aligned, which was possible prior to 69ba284b5e077075.)