Disable regeneration of regcharclass.h on win32
There's really no need to do this on the user side, as regcharclass.h is a real
file in the repository. Having this rule, while maybe handy for regexp
developers working on win32, otherwise only causes trouble if regcharclass.pl
changes in a way that doesn't require regcharclass.h to change.